// TypeScript/app/src/item/enums/item-name.enum.ts
export enum ItemName {
    AgedBrie = 'Aged Brie',
    Backstage = 'Backstage passes to a TAFKAL80ETC concert',
    SulfurasHandOfRagnaros = 'Sulfuras, Hand of Ragnaros',
    ConjuredManaCake = 'Conjured Mana Cake',
    DexterityVest = '+5 Dexterity Vest',
    ElixirOfTheMongoose = 'Elixir of the Mongoose'
}